+++ CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 1.12.39, 4/21/2020
+
+### Added
+
+Use RestEasy JAX-RS library
+
+### Removed
+
+For security reason, removed Jersey JAX-RS library
+
+### Changed
+
+Updated RestLoader to initialize RestEasy servlet dispatcher
+
+---
 ## Version 1.12.36, 4/16/2020
 
 ### Added